// fe/src/lib/api/generated/model/emailDeliveryItemResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface EmailDeliveryItemResponse {
  attempt_count: number;
  created_at: string;
  event_code: string;
  failure_code: string;
  id: string;
  last_attempt_at: string | null;
  next_attempt_at: string | null;
  sent_at: string | null;
  status: string;
  updated_at: string;
}